Medic and Tank are kinda working. Starting to work on specialist.

+++ First Game/Player.cpp
@@ -46,8 +46,7 @@ void Player::setBulletCount(int v)
 }
 
 
-
-float Player::getInitialHealth()
+int Player::getInitialHealth()
 {
 	return initialHealth;
 }
